Extract the declarations from the body
To allow declarations in the function being cached, the previous code wrapped
the body, which may containing declarations, with a LOCALLY form. This however
interferes with the implicit progn of the function being cached. Which causes
(RETURN-FROM fmt-wifi "no link") in FMT-WIFI to return NIL instead.
This was reported by Rafael Fernández[0].

A cleaner solution is to extract any declarations in the body and hoist them to
the beginning of DEFUN and replacing LOCALLY with PROGN

[0]: https://github.com/stumpwm/stumpwm-contrib/pull/61