Add pass-copy-menu
When selecting a password to copy via `pass-copy` I often don't
remember the name under which I stored the password. Autocomplete in
`pass-copy` helps a bit, but I think a bit more visibility would be
nice.

This PR adds the function `pass-copy-menu`, which displays the
available password entries in a menu. The user can then select or
narrow down an entry via autocompletion.