don't use sb-ext:process-close but rather detect our proper shutdown
Using sb-ext:process-close means that sbcl will not collect the status
change of the terminated process and leave the process as a zombie,
as it completely will stop caring about the process. A solution would
be to collect the exit status change ourselves by calling sb-impl::waitpid
but that one is not external and too much hassling with internals.

Rather I let the process exit be handled by sbcl and detect our
intentional process terminatino in radio-status-change by a variable.
Not too beautiful but perfectly okay as we want only one radio be
played at most all the time.