README cleanup
I really like having the readme in an org file. This change fixes some
source code blocks which were taged as example blocks, which broke
syntax highlighting of code fragments in the README. I have also
inserted spaces in various places, which aid readability; after all,
this is org mode, and we can collapse what we aren't focused upon, so
there is no benefit in having the file expand as one long block of
homogenous text.