reworked refreshing from stump, readme, presentations
Removed the patched stumpwm::eval-command function and instead hung a
refresh function on the event hook in stumpwm.

moved highlighting of current window/group into the present method.

added a lock to the redisplay to function to (hopefully) cut down on
freezes/hanging of the mode line.

rewrote goals and TODO portion of the readme.