Remove clim mode line
The clim mode line has been a failure. I added it here with rose-tinted glasses, thinking that others would see it and contribute to it. However my own motivation to make this a reality has died, and I think the best option at this point is to remove this from the repo, as it can cause confusion to users who expect a well functioning mode line based on clim as opposed to the current reality of a hackish mode line that is less functional than the default StumpWM mode line. 

I may continue working on this in the future, but until it is truly functional it should live in its own repo.