Fix bug with dangling timers
In certain cases, usually around loss of connection to the network, it
was possible for the MPD connection to be dropped while *mpd-timer*
was still active and scheduled. This led to mpd-ping being called
on the timer every *mpd-timeout* seconds, and uselessly popping up an
error message.

This fixes that bug by replacing the simple mpd-ping call in the timer
with a lambda that checks for *mpd-socket* and, if not connected,
cancels the timer. It also makes mpd-disconnect cancel the timer
outside the with-mpd-connection form, which means that in the case of
a dangling timer mpd-disconnect will still cancel it.