stumpish: Fix search for stumpwm PID if launched as stumpwm child process.
FreeBSD specific fix. In the FreeBSD the pgrep didn't list processes,
which are ancestors of pgrep even if they are matched with given
string. So, if stumpish was launched from the stumpwm itself, then it
couldn't find PID of stumpwm process.

This fix uses pgrep with '-a' key, which enable showing even the
ancestors of pgrep in the it's output.